refactor : VO 재설계
