+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/auth/AuthResult.kt
@@ -2,10 +2,8 @@ package com.whatever.caramel.core.domain.vo.auth
 
 import com.whatever.caramel.core.domain.vo.user.UserStatus
 
-data class UserAuth(
+data class AuthResult(
     val coupleId: Long? = null,
     val userStatus: UserStatus,
-    val nickname: String? = null,
-    val birthday: String? = null,
     val authToken: AuthToken,
 )
