+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/balanceGame/BalanceGameResult.kt
@@ -1,10 +1,9 @@
 package com.whatever.caramel.core.domain.vo.balanceGame
 
 import com.whatever.caramel.core.domain.entity.BalanceGame
-import com.whatever.caramel.core.domain.entity.BalanceGameOption
 
 data class BalanceGameResult(
     val gameInfo: BalanceGame,
-    val myChoice: BalanceGameOption?,
-    val partnerChoice: BalanceGameOption?,
+    val myChoice: BalanceGame.Option?,
+    val partnerChoice: BalanceGame.Option?,
 )
