--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/calendar/AnniversariesOnDate.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.core.domain.vo.calendar
-
-import com.whatever.caramel.core.domain.vo.couple.Anniversary
-import kotlinx.datetime.LocalDate
-
-data class AnniversariesOnDate(
-    val date: LocalDate,
-    val anniversaries: List<Anniversary>,
-)
