+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/calendar/Anniversary.kt
@@ -1,10 +1,10 @@
-package com.whatever.caramel.core.domain.vo.couple
+package com.whatever.caramel.core.domain.vo.calendar
 
+import com.whatever.caramel.core.domain.vo.calendar.AnniversaryType
 import kotlinx.datetime.LocalDate
 
 data class Anniversary(
     val type: AnniversaryType,
     val date: LocalDate,
     val label: String,
-    val isAdjustedForNonLeapYear: Boolean,
 )
