--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/calendar/HolidaysOnDate.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.core.domain.vo.calendar
-
-import com.whatever.caramel.core.domain.entity.Holiday
-import kotlinx.datetime.LocalDate
-
-data class HolidaysOnDate(
-    val date: LocalDate,
-    val holidays: List<Holiday>,
-)
