--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/calendar/ScheduleMetadata.kt
@@ -1,6 +1,0 @@
-package com.whatever.caramel.core.domain.vo.calendar
-
-data class ScheduleMetadata(
-    val contentId: Long,
-    val contentType: String,
-)
