+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/content/ContentData.kt
@@ -1,0 +1,8 @@
+package com.whatever.caramel.core.domain.vo.content
+
+data class ContentData(
+    val title: String,
+    val description: String,
+    val isCompleted: Boolean,
+    val contentAssignee: ContentAssignee,
+)
