+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/content/schedule/DateTimeInfo.kt
@@ -1,4 +1,4 @@
-package com.whatever.caramel.core.domain.vo.common
+package com.whatever.caramel.core.domain.vo.content.schedule
 
 data class DateTimeInfo(
     val startDateTime: String,
