--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/memo/MemoMetadata.kt
@@ -1,6 +1,0 @@
-package com.whatever.caramel.core.domain.vo.memo
-
-data class MemoMetadata(
-    val contentId: Long,
-    val contentType: String,
-)
