+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/user/UserAgreement.kt
@@ -1,0 +1,6 @@
+package com.whatever.caramel.core.domain.vo.user
+
+data class UserAgreement(
+    val serviceTerms: Boolean,
+    val privacyPolicy: Boolean,
+)
